+++ site/e2e/api.ts
@@ -1,4 +1,5 @@
 import type { Page } from "@playwright/test";
+import { expect } from "@playwright/test";
 import * as API from "api/api";
 import { coderPort } from "./constants";
 import { findSessionToken, randomName } from "./helpers";
@@ -47,3 +48,43 @@ export const createGroup = async (orgId: string) => {
   });
   return group;
 };
+
+export async function verifyConfigFlag(page: Page, flag: string) {
+  const config = await API.getDeploymentConfig();
+
+  const opt = config.options.find((option) => option.flag === flag);
+  if (opt === undefined) {
+    throw new Error(`Option with env ${flag} has undefined value.`);
+  }
+
+  // Map option type to test class name.
+  let type = "",
+    value = opt.value;
+  if (typeof value === "boolean") {
+    // Boolean options map to string (Enabled/Disabled).
+    type = value ? "option-enabled" : "option-disabled";
+    value = value ? "Enabled" : "Disabled";
+  } else if (typeof value === "number") {
+    type = "option-value-number";
+    value = String(value);
+  } else if (!value || value.length === 0) {
+    type = "option-value-empty";
+  } else if (typeof value === "string") {
+    type = "option-value-string";
+  } else if (typeof value === "object") {
+    type = "object-array";
+  } else {
+    type = "option-value-json";
+  }
+
+  // Special cases
+  if (opt.flag === "strict-transport-security" && opt.value === 0) {
+    type = "option-value-string";
+    value = "Disabled"; // Display "Disabled" instead of zero seconds.
+  }
+
+  const configOption = page.locator(
+    `div.options-table .option-${flag} .${type}`,
+  );
+  await expect(configOption).toHaveText(String(value));
+}
